refactored utility function out into a module
