+++ blog/index.html
@@ -26,7 +26,7 @@
     </template>
 
     <script src="../lib/showdown/showdown.js"></script>
-    <script src="blog.js"></script>
+    <script type="module" src="blog.js"></script>
 </body>
 
 </html>
