+++ lib/domutils.js
@@ -1,0 +1,10 @@
+// Function to wait for an element to exist
+export function waitForElement(selector, callback) {
+    const interval = setInterval(() => {
+        const element = document.querySelector(selector);
+        if (element) {
+            clearInterval(interval);
+            callback(element);
+        }
+    }, 100); // Check every 100ms
+}
